Factor out `CWuaGDialect` in favor of `CWuaDialect`
