+++ src/xrGame/vs2022/xrGame.vcxproj
@@ -341,7 +341,7 @@
     <ClInclude Include="..\..\xrServerEntities\script_dialects.h" />
     <ClInclude Include="..\..\xrServerEntities\script_dialect_lisp_macro.h" />
     <ClInclude Include="..\..\xrServerEntities\script_dialect_lua.h" />
-    <ClInclude Include="..\..\xrServerEntities\script_dialect_wua_g.h" />
+    <ClInclude Include="..\..\xrServerEntities\script_dialect_wua.h" />
     <ClInclude Include="..\..\xrServerEntities\script_engine.h" />
     <ClInclude Include="..\..\xrServerEntities\script_engine_export.h" />
     <ClInclude Include="..\..\xrServerEntities\script_engine_inline.h" />
@@ -356,7 +356,6 @@
     <ClInclude Include="..\..\xrServerEntities\script_ini_file_inline.h" />
     <ClInclude Include="..\..\xrServerEntities\script_dialect.h" />
     <ClInclude Include="..\..\xrServerEntities\script_dialect_lisp.h" />
-    <ClInclude Include="..\..\xrServerEntities\script_dialect_wua.h" />
     <ClInclude Include="..\..\xrServerEntities\script_lua_helper.h" />
     <ClInclude Include="..\..\xrServerEntities\script_net_packet.h" />
     <ClInclude Include="..\..\xrServerEntities\script_process.h" />
@@ -1974,7 +1973,7 @@
     <ClCompile Include="..\..\xrServerEntities\script_dialect_lisp_macro.cpp" />
     <ClCompile Include="..\..\xrServerEntities\script_dialect_lua.cpp" />
     <ClCompile Include="..\..\xrServerEntities\script_dialect_script.cpp" />
-    <ClCompile Include="..\..\xrServerEntities\script_dialect_wua_g.cpp" />
+    <ClCompile Include="..\..\xrServerEntities\script_dialect_wua.cpp" />
     <ClCompile Include="..\..\xrServerEntities\script_engine.cpp">
       <PrecompiledHeaderFile>pch_script.h</PrecompiledHeaderFile>
       <PrecompiledHeaderOutputFile>$(IntDir)$(ProjectName)_script.pch</PrecompiledHeaderOutputFile>
@@ -2010,7 +2009,6 @@
     </ClCompile>
     <ClCompile Include="..\..\xrServerEntities\script_dialect.cpp" />
     <ClCompile Include="..\..\xrServerEntities\script_dialect_lisp.cpp" />
-    <ClCompile Include="..\..\xrServerEntities\script_dialect_wua.cpp" />
     <ClCompile Include="..\..\xrServerEntities\script_lua_helper.cpp">
       <PrecompiledHeaderFile>pch_script.h</PrecompiledHeaderFile>
       <PrecompiledHeaderOutputFile>$(IntDir)$(ProjectName)_script.pch</PrecompiledHeaderOutputFile>
